add invoke typings and seperate mouse detection logic
